Move Jackson to separate class
