+++ core/src/main/java/com/sap/ai/sdk/core/AiCoreService.java
@@ -1,14 +1,8 @@
 package com.sap.ai.sdk.core;
 
-import static com.fasterxml.jackson.databind.MapperFeature.DEFAULT_VIEW_INCLUSION;
+import static com.sap.ai.sdk.core.JacksonConfiguration.getDefaultObjectMapper;
 
-import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.PropertyAccessor;
-import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.json.JsonMapper;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.fasterxml.jackson.module.paramnames.ParameterNamesModule;
 import com.google.common.annotations.Beta;
 import com.google.common.collect.Iterables;
 import com.sap.cloud.sdk.cloudplatform.connectivity.ApacheHttpClient5Accessor;
@@ -236,30 +230,5 @@ public class AiCoreService {
           .property(RESOURCE_GROUP_HEADER_PROPERTY, resourceGroup)
           .build();
     }
-  }
-
-  /**
-   * Default object mapper used for JSON de-/serialization. <b>Only intended for internal usage
-   * within this SDK</b>. Largely follows the defaults set by Spring.
-   *
-   * @return A new object mapper with the default configuration.
-   * @see <a
-   *     href="https://docs.spring.io/spring-framework/docs/current/javadoc-api/org/springframework/http/converter/json/Jackson2ObjectMapperBuilder.html">Jackson2ObjectMapperBuilder</a>
-   */
-  @Nonnull
-  @Beta
-  public static JsonMapper getDefaultObjectMapper() {
-    return JsonMapper.builder()
-        .addModule(new JavaTimeModule())
-        .addModule(new ParameterNamesModule())
-        // Disable automatic detection of getters and setters
-        .visibility(PropertyAccessor.GETTER, JsonAutoDetect.Visibility.NONE)
-        .visibility(PropertyAccessor.SETTER, JsonAutoDetect.Visibility.NONE)
-        .serializationInclusion(JsonInclude.Include.NON_NULL)
-        // for generated code unknown properties should always be stored as custom fields
-        // still added for any non-generated code and additional safety
-        .disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES)
-        .disable(DEFAULT_VIEW_INCLUSION)
-        .build();
   }
 }
